requirements.txt
flask
PyGithub
python-dotenv
PyYAML
google-generativeai
